Move builders to separate subclass
