+++ hudi-spark-datasource/hudi-spark-common/src/main/scala/org/apache/hudi/Iterators.scala
@@ -408,7 +408,7 @@ object LogFileIterator extends SparkAdapterSupport {
       logRecordScannerBuilder.withRecordMerger(
         HoodieRecordUtils.createRecordMerger(tableState.tablePath, EngineType.SPARK, tableState.recordMergerImpls.asJava, tableState.recordMergerStrategy))
 
-      val scanner = logRecordScannerBuilder.build().asInstanceOf[HoodieMergedLogRecordScanner]
+      val scanner = logRecordScannerBuilder.build()
 
       closing(scanner) {
         // NOTE: We have to copy record-map (by default immutable copy is exposed)
